+++ central/src/temperature.c
@@ -9,13 +9,13 @@ void* main_temperature(){
     while (1)
     {
         char response[16];
-        float temperature, humidity;
+        double temperature, humidity;
         char res;
         
         request_BME280_data(response);
         printf("response: %s\n", response);
-        sscanf(response, "%1c %f %f", res, temperature, humidity);
-        printf("UMIDADE: %f\n", humidity);
+        sscanf(response, "%1c %lf %lf", res, temperature, humidity);
+        printf("UMIDADE: %lf\n", humidity);
         print_temperature(temperature, humidity);
 
         usleep(1000000);
